Introduce `ids` to the source data
The rationale for this is that since this component now offers the
ability to edit/create new elements, it makes sense to uniquely
identify each of the elements in the chart. Otherwise it becomes really
hard for the client to figure out which elements were changed in the
chart based on the output of the callbacks only.